[fix/feat] Ajuste no cadastro de consultas

+++ src/main/java/com/digitalhouse/clinicaodontologicag6/service/IClinicaService.java
@@ -3,10 +3,11 @@ package com.digitalhouse.clinicaodontologicag6.service;
 import java.util.List;
 
 public interface IClinicaService<T> {
+
     T create(T t);
-    T getById(int id);
+    T getById(Long id);
     List<T> getAll();
-    String delete(int id);
-    T update(T t, int id);
+    String delete(Long id);
+    T update(T t, Long id);
 
 }
